+++ src/constants/ReservedIdentifierNames.ts
@@ -1,0 +1,11 @@
+export const reservedIdentifierNames = [
+    // reserved identifiers
+    'byte', 'case', 'char', 'do', 'else', 'enum', 'eval', 'for', 'goto',
+    'if', 'in', 'int', 'let', 'long', 'new', 'null', 'this', 'true', 'try',
+    'var', 'void', 'with',
+
+    // reserved global object identifiers
+    'Array', 'Attr', 'Audio', 'Blob', 'Cache', 'Date', 'Error', 'Event',
+    'Feed', 'File', 'Hz', 'Image', 'Intl', 'Lock', 'Map', 'Math', 'Node',
+    'Proxy', 'Range', 'Rect', 'Set', 'Table', 'Text', 'Touch'
+];