+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Change Log
 
+v3.2.3
+---
+* Fixed missing transformation of string array calls in some cases
+* Fixed generation of reserved identifier names like `Map` or `Set` for `mangled` and `mangled-shuffled` identifier names generators
+
 v3.2.2
 ---
 * Fixed https://github.com/javascript-obfuscator/javascript-obfuscator/issues/1039